requirements.txt
numpy==1.23.3
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.1
transformers==4.37.1
pandas==2.2.1
pillow==10.2.0
scikit-learn==1.4.0
wandb==0.19.2
